+++ src/main/java/org/starchartlabs/lure/command/PushCommand.java
@@ -39,7 +39,6 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 
-//TODO romeara test
 /**
  * Command line sub-command that will push raw data to a web URL following patterns defined for GitHub webhook calls
  *