+++ CHANGE_LOG.md
@@ -5,3 +5,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added "push" command which allows sending pre-defined event payloads to a URL in the same manner as GitHub webhooks